simulation/stage17_controlled_rebuild/requirements.txt
numpy
scipy
matplotlib
pandas
PyYAML
